csfix

+++ src/Controller/UserController.php
@@ -8,13 +8,13 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Form\Type\UserType;
 use App\Service\UserServiceInterface;
-use Doctrine\DBAL\Exception\ForeignKeyConstraintViolationException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
+use Symfony\Component\Form\Extension\Core\Type\FormType;
+use Doctrine\DBAL\Exception\ForeignKeyConstraintViolationException;
 
 /**
  * Class UserController.
@@ -37,14 +37,9 @@ class UserController extends AbstractController
         if ($this->isGranted('ROLE_ADMIN')) {
             $page = $request->query->getInt('page', 1);
             $pagination = $this->userService->getPaginatedList($page);
-
-            $result = $this->render('user/index.html.twig', [
-                'pagination' => $pagination]);
+            $result = $this->render('user/index.html.twig', ['pagination' => $pagination]);
         } else {
-            $this->addFlash(
-                'danger',
-                $this->translator->trans('message.access_denied')
-            );
+            $this->addFlash('danger', $this->translator->trans('message.access_denied'));
             $result = $this->redirectToRoute('recipe_index');
         }
 
@@ -61,22 +56,14 @@ class UserController extends AbstractController
 
             if ($form->isSubmitted() && $form->isValid()) {
                 $this->userService->save($user);
-
-                $this->addFlash(
-                    'success',
-                    $this->translator->trans('message.created_successfully')
-                );
+                $this->addFlash('success', $this->translator->trans('message.created_successfully'));
 
                 return $this->redirectToRoute('user_index');
             }
 
-            return $this->render('user/create.html.twig', [
-                'form' => $form->createView()]);
+            return $this->render('user/create.html.twig', ['form' => $form->createView()]);
         } else {
-            $this->addFlash(
-                'danger',
-                $this->translator->trans('message.access_denied')
-            );
+            $this->addFlash('danger', $this->translator->trans('message.access_denied'));
             $result = $this->redirectToRoute('recipe_index');
         }
 
@@ -86,14 +73,7 @@ class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'user_edit', requirements: ['id' => '\d+'], methods: ['GET|PUT'])]
     public function edit(Request $request, User $user): Response
     {
-        $form = $this->createForm(
-            UserType::class,
-            $user,
-            [
-                'method' => 'PUT',
-                'action' => $this->generateUrl('user_edit', ['id' => $user->getId()]),
-            ]
-        );
+        $form = $this->createForm(UserType::class, $user, ['method' => 'PUT', 'action' => $this->generateUrl('user_edit', ['id' => $user->getId()])]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -107,23 +87,16 @@ class UserController extends AbstractController
             return $this->redirectToRoute('user_index');
         }
 
-        return $this->render('user/edit.html.twig', [
-            'form' => $form->createView(),
-            'user' => $user,
-        ]);
+        return $this->render('user/edit.html.twig', ['form' => $form->createView(), 'user' => $user]);
     }
 
     #[Route('/{id}', name: 'user_show', requirements: ['id' => '\d+'], methods: ['GET'])]
     public function show(User $user): Response
     {
         if ($this->isGranted('ROLE_ADMIN')) {
-            $result = $this->render('user/show.html.twig', [
-                'user' => $user]);
+            $result = $this->render('user/show.html.twig', ['user' => $user]);
         } else {
-            $this->addFlash(
-                'danger',
-                $this->translator->trans('message.access_denied')
-            );
+            $this->addFlash('danger', $this->translator->trans('message.access_denied'));
             $result = $this->redirectToRoute('recipe_index');
         }
 
@@ -133,6 +106,12 @@ class UserController extends AbstractController
     #[Route('/{id}/delete', name: 'user_delete', requirements: ['id' => '\d+'], methods: ['DELETE|GET'])]
     public function delete(Request $request, User $user): Response
     {
+        if ($this->isGranted('ROLE_USER') && $this->getUser() === $user) {
+            $this->addFlash('danger', $this->translator->trans('message.cannot_delete_user'));
+
+            return $this->redirectToRoute('user_index');
+        }
+
         $form = $this->createForm(FormType::class, $user, [
             'method' => 'DELETE',
             'action' => $this->generateUrl('user_delete', ['id' => $user->getId()]),
